--- docs/CHANGELOG.md
@@ -1,24 +1,0 @@
-# Changelog
-
-# Javascript library for rendering standard music notation in a browser.
-
-# Fork of abcjs Version 6.2.2
-
-## [0.0.2] - 2023-08-11
-
-## Added
-
-- Added parse-music.js to src/const as common source of re-used values
-
-## Changed
-
-- Updated abc_parse_music.js to import constants from parse_music
-- Updated abc_parser_lint.js to import constants from parse_music
-- Updated decoration.js to handle abbreviated glissando forms
-- Modified glissando.drawGlissando to handle grace note case
-
-## [0.0.1] - 2023-08-11
-
-## Added
-
-- Initial commit
